+++ database/migrations/2025_08_20_101317_create_shipment_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('shipment_id')->constrained('shipments')->onDelete('cascade');
             $table->foreignId('production_batch_id')->constrained('production_batches')->onDelete('cascade');
-            $table->decimal('qty_shipped', 10, 3);
+            $table->integer('qty_shipped');
             $table->string('uom')->default('cartons');
             $table->text('notes')->nullable();
             $table->timestamps();
